Added OrderedDefaultDict to reduce checks on dict and errors
When we know that the dict will have values as list,
it is beneficial to use the defaultdict datatype,
we don't have to put an extra code of checks and
validations thereby reducing the code and errors.

And to maintain the order of the deployment on Nulecule artifacts
we need OrderedDict. So to merge these two kinds of dict together
a new class has been added named OrderedDefaultDict.