Artifacts should be specified for the provider
For a provider mentioned by the user, artifacts should be
present in Nulecule file. Earlier if a provider was mentioned
and even if that provider's artifacts were not specified in
Nulecule file, Atomic App would go ahead. Now it fails error out
if artifacts for that provider are not mentioned in Nulecule file.

Fixes issue #641