Atomic App will not create redundant folder on `atomicapp stop`
Previously when `atomicapp stop projectatomic/helloapache` was
called atomicapp would create a redundant folder in
`/var/lib/atomicapp` and fail saying that `answers` file not
found. Though it was failing but the folder would still remain
there, now there are checks added to see if user has provided
the image name and the action is stop, the at the location where
it creates a directory, the new checks will cause it to fail.

Fixes issue #683