Added cert param in `make_rest_request` to pass client side certs
Added cert param so that now we can pass client side certs
while making request to API server, this can be passed as
path to file that contains cert and key or it can be a tuple
which has path ti cert and key.

This is part of the effort to add support to authenticate to
server via cert in addition to access_token.